Extract base class `Service`
