+++ src/main/php/web/Service.class.php
@@ -1,0 +1,37 @@
+<?php namespace web;
+
+use lang\Value;
+
+abstract class Service implements Value {
+  protected $environment;
+
+  /**
+   * Creates a new web application inside a given environment
+   *
+   * @param  web.Environment $environment
+   */
+  public function __construct(Environment $environment) {
+    $this->environment= $environment;
+  }
+
+  /** @return web.Environment */
+  public function environment() { return $this->environment; }
+
+  public abstract function serve($server, $environment);
+
+  /** @return string */
+  public function toString() { return nameof($this).'('.$this->environment->docroot().')'; }
+
+  /** @return string */
+  public function hashCode() { return spl_object_hash($this); }
+
+  /**
+   * Comparison
+   *
+   * @param  var $value
+   * @return int
+   */
+  public function compareTo($value) {
+    return $value === $this ? 0 : 1;
+  }
+}
